新增資料表設定

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -11,22 +11,28 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('account')->nullable()->comment('帳號名稱');
-            $table->string('name')->nullable()->comment('名字');
-            $table->string('email')->nullable()->comment('電子郵件');
-            $table->string('nation')->nullable()->comment('國際碼');
-            $table->string('mobile')->nullable()->comment('行動電話');
-            $table->string('phone')->nullable()->comment('國際碼+行動電話');
-            $table->string('password')->comment('名字');
-            $table->string('address')->nullable()->comment('地址');
-            $table->date('birthDay')->nullable()->comment('生日');
-            $table->timestamp('email_verified_at')->nullable();
-            $table->boolean('status')->default(0)->comment('-9:已刪除, -1:停用, 0:未完成註冊, 1:完成註冊');
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (env('DB_MIGRATE_USERS')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('account')->nullable()->comment('帳號名稱');
+                $table->string('name')->nullable()->comment('名字');
+                $table->string('email')->nullable()->comment('電子郵件');
+                $table->string('nation')->nullable()->comment('國際碼');
+                $table->string('mobile')->nullable()->comment('行動電話');
+                $table->string('phone')->nullable()->comment('國際碼+行動電話');
+                $table->string('password')->comment('名字');
+                $table->string('address')->nullable()->comment('地址');
+                $table->string('verify_mode')->nullable()->comment('兩階段驗證方式');
+                $table->string('2fa_secret')->nullable()->comment('2fa密鑰');
+                $table->string('otp_code',10)->nullable()->comment('簡訊碼');
+                $table->datetime('otp_time')->nullable()->comment('簡訊碼到期時間');
+                $table->date('birthDay')->nullable()->comment('生日');
+                $table->boolean('lock_on')->default(0)->comment('0:未鎖定, 超過次數限制鎖定');
+                $table->boolean('status')->default(0)->comment('-9:已刪除, -3:鎖定, -1:停用, 0:未完成註冊, 1:完成註冊');
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -34,6 +40,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if (env('DB_MIGRATE_USERS')) {
+            Schema::dropIfExists('users');
+        }
     }
 };
